refactor: keep Eidoverse signal helpers independent of source reads (#7105)
