+++ server/lib/index.js
@@ -537,6 +537,7 @@ export * from './domainAutonomy.js';
 export * from './domainBudgets.js';
 export * from './eidoverseWorldDesign.js';
 export * from './eidoverseWorldLabels.js';
+export * from './eidoverseWorldSignals.js';
 export * from './errorHandler.js';
 export * from './extensionErrors.js';
 export * from './fetchErrorChain.js';
